+++ config/app/index.ts
@@ -1,8 +1,9 @@
 import { env } from "../env";
-import { AppConfig } from "./interface";
+import type { AppConfig, AppDeployment } from "./interface";
 
-const { APP_PORT } = env;
+const { APP_DEPLOYMENT, APP_PORT } = env;
 
 export const appConfig: AppConfig = {
+  deployment: (APP_DEPLOYMENT as AppDeployment) || "development",
   port: APP_PORT || "5000",
 };