+++ config/app/interface.ts
@@ -1,3 +1,6 @@
+export type AppDeployment = "development" | "production";
+
 export interface AppConfig {
+  deployment: AppDeployment;
   port: string;
 }