+++ controller/service/interface.ts
@@ -1,6 +1,6 @@
 import { Controller } from "../interface";
 
 export interface ServiceController {
-  postService: Controller;
-  getService: Controller;
+  post: Controller;
+  get: Controller;
 }