+++ repository/index.ts
@@ -1,1 +1,2 @@
 export { createServiceRepository } from "./service";
+export { createServiceInstanceRepository } from "./serviceInstance";