+++ repository/service/index.ts
@@ -1,21 +1,28 @@
 import { prisma } from "@/service";
 import type {
   ServiceRepository,
-  ServiceCreateInput,
-  Service,
+  CreateOne,
+  CreateOneInput,
+  ReadAll,
+  UpdateOne,
+  DeleteOne,
+  UpdateOneInput,
 } from "./interface";
 
 const { service } = prisma;
 
-const readAll = async (): Promise<Service[]> => {
-  try {
-    return await service.findMany();
-  } catch (error) {
-    throw error;
-  }
+export const createServiceRepository = (): ServiceRepository => {
+  const serviceRepository: ServiceRepository = {
+    createOne,
+    readAll,
+    updateOne,
+    deleteOne,
+  };
+
+  return serviceRepository;
 };
 
-const createOne = async (s: ServiceCreateInput): Promise<Service> => {
+const createOne: CreateOne = async (s: CreateOneInput) => {
   try {
     return await service.create({ data: s });
   } catch (error) {
@@ -23,11 +30,26 @@ const createOne = async (s: ServiceCreateInput): Promise<Service> => {
   }
 };
 
-export const createServiceRepository = (): ServiceRepository => {
-  const serviceRepository: ServiceRepository = {
-    createOne,
-    readAll,
-  };
+const readAll: ReadAll = async () => {
+  try {
+    return await service.findMany();
+  } catch (error) {
+    throw error;
+  }
+};
 
-  return serviceRepository;
+const updateOne: UpdateOne = async (id: number, p: UpdateOneInput) => {
+  try {
+    return await service.update({ where: { id }, data: p });
+  } catch (error) {
+    throw error;
+  }
+};
+
+const deleteOne: DeleteOne = async (id: number) => {
+  try {
+    return service.delete({ where: { id } });
+  } catch (error) {
+    throw error;
+  }
 };