+++ repository/service/interface.ts
@@ -1,9 +1,16 @@
 import { Prisma, Service } from "@prisma/client";
 
 export interface ServiceRepository {
-  createOne: (s: Prisma.ServiceCreateInput) => Promise<Service>;
-  readAll: () => Promise<Service[]>;
+  createOne: CreateOne;
+  readAll: ReadAll;
+  updateOne: UpdateOne;
+  deleteOne: DeleteOne;
 }
 
-export type { Service } from "@prisma/client";
-export type ServiceCreateInput = Prisma.ServiceCreateInput;
+export type { Service };
+export type CreateOne = (p: CreateOneInput) => Promise<Service>;
+export type CreateOneInput = Prisma.ServiceCreateInput;
+export type ReadAll = () => Promise<Service[]>;
+export type UpdateOne = (id: number, p: UpdateOneInput) => Promise<Service>;
+export type UpdateOneInput = Prisma.ServiceUpdateInput;
+export type DeleteOne = (id: number) => Promise<Service>;