+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.0.2] -
+
+### Changed
+
+- name of controller, repository, api's type and interface
+
+### Added
+
+- serviceInstance CRUD
+- module alias
+- app config DEPLOYMENT
+
 ## [0.0.1] - 2024-03-14
 
 ### Added